Extract ThumbnailPreviewStrip to be reusable
